mybackend/requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
ipython
django-extensions
django-environ